new module for InteractiveBrowserCredential tests
